+++ apps/filecoin-site/src/app/(homepage)/data/communityLinks.ts
@@ -20,17 +20,17 @@ export const communityLinks = [
   },
   {
     title: 'Attend a Filecoin event',
-    href: FILECOIN_FOUNDATION_URLS.events,
+    href: FILECOIN_FOUNDATION_URLS.events.href,
     icon: MapPinIcon,
   },
   {
     title: 'Get involved on GitHub',
-    href: FILECOIN_URLS.github,
+    href: FILECOIN_URLS.github.href,
     icon: GithubLogoIcon,
   },
   {
     title: 'Become a Filecoin Orbit ambassador',
-    href: FILECOIN_FOUNDATION_URLS.orbit,
+    href: FILECOIN_FOUNDATION_URLS.orbit.href,
     icon: MegaphoneIcon,
   },
 ] as const satisfies Array<LinkCardData>
