+++ apps/filecoin-site/src/app/_components/Footer/LegalSection.tsx
@@ -1,14 +1,15 @@
 import { InternalTextLink } from '@filecoin-foundation/ui/TextLink/InternalTextLink'
 
-import { legalLinks } from '@/constants/navigation'
+import { FILECOIN_FOUNDATION_URL } from '@/constants/siteMetadata'
 
 import FilecoinFoundationLogo from '@/assets/logos/filecoin-foundation-logo.svg'
 
+import { legalLinks } from '../Navigation/constants/navigation'
 export function LegalSection() {
   return (
     <div className="flex flex-col-reverse items-center justify-between gap-8 py-16 sm:flex-row md:gap-16">
       <a
-        href="https://fil.org/"
+        href={FILECOIN_FOUNDATION_URL}
         className="focus:brand-outline flex items-center gap-2 text-xs text-[var(--color-paragraph-text)] hover:text-white"
         aria-label="Managed by Filecoin Foundation &ndash; Visit website"
       >
