+++ apps/filecoin-site/src/app/_components/Navigation/NavigationMenu/NavigationMenuPanel.tsx
@@ -2,21 +2,19 @@
 
 import { clsx } from 'clsx'
 
-import { type ExpandedNavItem } from '@/constants/navigation'
-
 import { useBackgroundVariant } from '@/components/BackgroundVariantContext'
 import { backgroundVariants } from '@/components/Section'
 
-import { variantMapping } from './constants'
+import { type LabelledExpandedNavItems } from '../constants/navigation'
+import { variantMapping } from '../constants/variantMapping'
+
 import { NavigationMenuLink } from './NavigationMenuLink'
 
-type NavigationMenuPanelProps = {
-  items: ExpandedNavItem['items']
-}
+const roundedStyle = 'rounded-2xl'
 
-const roundedStyle = 'rounded-xl'
-
-export function NavigationMenuPanel({ items }: NavigationMenuPanelProps) {
+export function NavigationMenuPanel({
+  items,
+}: Pick<LabelledExpandedNavItems, 'items'>) {
   const backgroundVariant = useBackgroundVariant()
   const desktopBackgroundVariant = variantMapping[backgroundVariant]
 
@@ -39,7 +37,12 @@ export function NavigationMenuPanel({ items }: NavigationMenuPanelProps) {
               <p className="mb-2 px-4 text-sm text-[var(--color-paragraph-text)]">
                 {item.title}
               </p>
-              <ul className="grid grid-flow-col grid-rows-2 gap-2">
+              <ul
+                className="grid grid-rows-2 gap-2"
+                style={{
+                  gridTemplateColumns: `repeat(${Math.floor(item.links.length / 2)}, minmax(0, 1fr))`,
+                }}
+              >
                 {item.links.map((link) => (
                   <li key={link.href}>
                     <NavigationMenuLink {...link} />
