--- apps/filecoin-site/src/app/_components/Navigation/constants.ts
@@ -1,26 +1,0 @@
-import { PATHS } from '@/constants/paths'
-
-import type { SectionProps } from '../Section'
-
-export const NAV_LINKS = [
-  PATHS.LEARN,
-  PATHS.CASE_STUDIES,
-  PATHS.STORE_DATA,
-  PATHS.PROVIDE_STORAGE,
-  PATHS.BUILD_ON_FILECOIN,
-  PATHS.COMMUNITY_HUB,
-  PATHS.BLOG,
-] as const
-
-type VariantMapping = Record<
-  SectionProps['backgroundVariant'],
-  'light' | 'dark'
->
-
-export const variantMapping: VariantMapping = {
-  light: 'light',
-  gray: 'light',
-  transparent: 'light',
-  dark: 'dark',
-  transparentDark: 'dark',
-}
