+++ apps/filecoin-site/src/app/_components/Navigation/hooks/useIsNavigationMenuActive.ts
@@ -2,14 +2,15 @@ import { usePathname } from 'next/navigation'
 
 import { isInternalLink } from '@filecoin-foundation/utils/linkUtils'
 
-import type { ExpandedNavItem } from '@/constants/navigation'
 import { BASE_DOMAIN } from '@/constants/siteMetadata'
 
-export function useIsNavigationMenuActive(items: ExpandedNavItem['items']) {
+import type { LabelledExpandedNavItems } from '../constants/navigation'
+
+export function useIsNavigationMenuActive(
+  items: LabelledExpandedNavItems['items'],
+) {
   const pathname = usePathname()
-
   const links = items.flatMap((item) => item.links).map((item) => item.href)
-
   const internalLinks = links.filter((item) => {
     return isInternalLink(item, BASE_DOMAIN)
   })
