--- apps/filecoin-site/src/app/_constants/navigation.ts
@@ -1,214 +1,0 @@
-import { PATHS } from './paths'
-import {
-  FILECOIN_DOCS_URL,
-  FILECOIN_DOCS_URLS,
-  FILECOIN_FOUNDATION_URLS,
-  FILECOIN_URLS,
-} from './siteMetadata'
-
-export type NavItem = { label: string; href: string }
-
-type FooterNavigationItem = { title: string; items: Array<NavItem> }
-
-export type ExpandedNavItem = {
-  label: NavItem['label']
-  items: Array<{
-    title: string
-    links: Array<NavItem & { description: string }>
-  }>
-}
-
-const { github, securityBugBounty } = FILECOIN_URLS
-const { events, governance, grants, orbit } = FILECOIN_FOUNDATION_URLS
-
-const blockExplorersLinks: Array<NavItem> = [
-  { label: 'Beryx', href: 'https://beryx.io/' },
-  { label: 'Blockscout (FEVM)', href: 'https://www.blockscout.com/' },
-  { label: 'Filfox', href: 'https://filfox.info/' },
-  { label: 'Filscan', href: 'https://filscan.io/en/' },
-]
-
-const communityLinks: Array<NavItem> = [
-  { label: PATHS.COMMUNITY_HUB.label, href: PATHS.COMMUNITY_HUB.path },
-  { label: 'Events', href: events },
-  { label: 'Orbit', href: orbit },
-  { label: 'FIPs & Governance', href: governance },
-]
-
-export const legalLinks: Array<NavItem> = [
-  { label: PATHS.PRIVACY_POLICY.label, href: PATHS.PRIVACY_POLICY.path },
-  { label: PATHS.TERMS_OF_USE.label, href: PATHS.TERMS_OF_USE.path },
-]
-
-export const navigationBis: Array<NavItem | ExpandedNavItem> = [
-  {
-    label: PATHS.LEARN.label,
-    items: [
-      {
-        title: 'Understand Filecoin',
-        links: [
-          {
-            label: PATHS.LEARN.label,
-            description: 'Understand what Filecoin is and how it works',
-            href: PATHS.LEARN.path,
-          },
-          {
-            label: 'Case Studies',
-            description: 'See how others are using Filecoin in the real world',
-            href: PATHS.CASE_STUDIES.path,
-          },
-        ],
-      },
-    ],
-  },
-  { label: PATHS.STORE_DATA.label, href: PATHS.STORE_DATA.path },
-  { label: PATHS.PROVIDE_STORAGE.label, href: PATHS.PROVIDE_STORAGE.path },
-  {
-    label: 'Developers',
-    items: [
-      {
-        title: 'Explore',
-        links: [
-          {
-            label: 'Build on Filecoin',
-            description: 'Start building apps on the Filecoin network',
-            href: PATHS.BUILD_ON_FILECOIN.path,
-          },
-          {
-            label: 'Documentation',
-            description: 'Official documentation for Filecoin',
-            href: FILECOIN_DOCS_URL,
-          },
-          {
-            label: 'Cookbook',
-            description: 'Recipes for building with Filecoin and the FVM',
-            href: FILECOIN_DOCS_URLS.builderCookbook,
-          },
-          {
-            label: 'GitHub',
-            description: "Explore Filecoin's open-source repositories",
-            href: FILECOIN_URLS.github,
-          },
-        ],
-      },
-      {
-        title: 'Contribute',
-        links: [
-          {
-            label: 'Grants',
-            description: 'Funding opportunities to build in the ecosystem',
-            href: grants,
-          },
-          {
-            label: 'Bug Bounty',
-            description: 'Help find vulnerabilities and get rewarded',
-            href: securityBugBounty,
-          },
-        ],
-      },
-    ],
-  },
-  {
-    label: 'Network',
-    items: [
-      {
-        title: 'Block Explorers',
-        links: [
-          {
-            label: 'Beryx',
-            description: 'Explorer and API for accessing real-time data',
-            href: 'https://beryx.io/',
-          },
-          {
-            label: 'Blockscout (FEVM)',
-            description: 'Open-source explorer for FEVM smart contracts',
-            href: 'https://www.blockscout.com/',
-          },
-          {
-            label: 'Filfox',
-            description: 'Explorer with rich network and storage statistics',
-            href: 'https://filfox.info/',
-          },
-          {
-            label: 'Filscan',
-            description: 'Explorer for blocks, deals, and addresses',
-            href: 'https://filscan.io/en/',
-          },
-        ],
-      },
-      {
-        title: 'Network Monitoring',
-        links: [
-          {
-            label: 'Network Status',
-            description: 'The status of the Filecoin networks',
-            href: 'https://status.filecoin.io/',
-          },
-          {
-            label: 'Network Health',
-            description: 'Filecoin chain activity and performance',
-            href: 'https://dashboard.starboard.ventures/',
-          },
-        ],
-      },
-    ],
-  },
-  {
-    label: 'Community',
-    items: [
-      {
-        title: 'Get involved',
-        links: [
-          {
-            label: PATHS.COMMUNITY_HUB.label,
-            description: 'Explore ways to contribute to the ecosystem',
-            href: PATHS.COMMUNITY_HUB.path,
-          },
-          {
-            label: 'Events',
-            description: 'Join meetups, hackathons, and conferences',
-            href: events,
-          },
-          {
-            label: 'Orbit',
-            description: 'Become a Filecoin Orbit ambassador',
-            href: orbit,
-          },
-          {
-            label: 'FIPs & Governance',
-            description: 'Propose, discuss, and shape future upgrades',
-            href: governance,
-          },
-        ],
-      },
-    ],
-  },
-  { label: PATHS.BLOG.label, href: PATHS.BLOG.path },
-]
-
-const navigation: Array<NavItem> = [
-  { label: PATHS.LEARN.label, href: PATHS.LEARN.path },
-  { label: PATHS.CASE_STUDIES.label, href: PATHS.CASE_STUDIES.path },
-  { label: PATHS.STORE_DATA.label, href: PATHS.STORE_DATA.path },
-  { label: PATHS.PROVIDE_STORAGE.label, href: PATHS.PROVIDE_STORAGE.path },
-  { label: PATHS.BUILD_ON_FILECOIN.label, href: PATHS.BUILD_ON_FILECOIN.path },
-  { label: PATHS.BLOG.label, href: PATHS.BLOG.path },
-]
-
-const resources: Array<NavItem> = [
-  { label: 'Documentation', href: FILECOIN_DOCS_URL },
-  { label: 'Cookbook', href: FILECOIN_DOCS_URLS.builderCookbook },
-  { label: 'GitHub', href: github },
-  { label: 'Grants', href: grants },
-  { label: 'Bug Bounty', href: securityBugBounty },
-  { label: 'Brand Kit', href: 'https://hub.fil.org/design' },
-  { label: 'Network Status', href: 'https://status.filecoin.io/' },
-  { label: 'Network Health', href: 'https://dashboard.starboard.ventures/' },
-]
-
-export const footerNavigationItems: Array<FooterNavigationItem> = [
-  { title: 'Navigation', items: navigation },
-  { title: 'Resources', items: resources },
-  { title: 'Block Explorers', items: blockExplorersLinks },
-  { title: 'Community', items: communityLinks },
-]
