+++ apps/filecoin-site/src/app/community-hub/data/ecosystemResources.ts
@@ -1,3 +1,5 @@
+import { FILECOIN_FOUNDATION_URL } from '@/constants/siteMetadata'
+
 import ANSALogo from '@/assets/logos/ansa-logo.svg'
 import FIDLLogo from '@/assets/logos/fidl-logo.png'
 import FilBLogo from '@/assets/logos/fil-b-logo.png'
@@ -20,7 +22,7 @@ export const ecosystemResources = [
       color: '#fff',
     },
     cta: {
-      href: 'https://fil.org/',
+      href: FILECOIN_FOUNDATION_URL,
       text: CTA_TEXT,
     },
   },
